extracted MissionDetailsIconList and MissionDetailsButton into their own files
