+++ src/app/component/MissionDetailsCard/MissionDetailsCard.jsx
@@ -1,51 +1,41 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import { H5, Body2 } from ".";
-import UserPhoneUnverifiedPopup from "./UserPhoneUnverifiedPopup";
+import { H5, Body2 } from "../";
 import {
   Divider,
-  Button,
-  Grid,
   Box,
-  Avatar,
   Card,
   CardHeader,
   CardContent,
   CardMedia,
   CardActions,
 } from "@material-ui/core";
-import { color } from "../../theme";
+
+import UserPhoneUnverifiedPopup from "../UserPhoneUnverifiedPopup";
+import MissionDetailsIconList from "./MissionDetailsIconList";
+import { MissionDetailsButton, MissionDetailsUnassignMeButton } from "./MissionDetailsButton";
+
 import PersonIcon from "@material-ui/icons/Person";
 import LocationOnIcon from "@material-ui/icons/LocationOn";
 import ScheduleIcon from "@material-ui/icons/Schedule";
-
 import AttachMoneyIcon from "@material-ui/icons/AttachMoney";
-import cameraImage from "../../img/placeholderBackground.svg";
+import cameraImage from "../../../img/placeholderBackground.svg";
 
 import { makeStyles } from "@material-ui/core/styles";
-import styled from "styled-components";
 
-import { MissionStatus, MissionFundedStatus } from "../model";
-
-const StyledButton = styled(Button)`
-  flex-grow: 1;
-`;
+import { MissionStatus, MissionFundedStatus } from "../../model";
 
 const useStyles = makeStyles((theme) => ({
   cardHeader: {
     paddingBottom: theme.spacing(1),
   },
+  cardAction: {
+    paddingLeft: theme.spacing(2),
+    paddingRight: theme.spacing(2),
+  },
   subheader: {
     marginTop: theme.spacing(0.3),
-  },
-  contentTypography: {
-    paddingLeft: theme.spacing(0.5),
-  },
-  avatar: {
-    height: "25px",
-    width: "25px",
-    marginBottom: theme.spacing(0.5),
   },
   image: {
     height: 0,
@@ -60,10 +50,6 @@ const useStyles = makeStyles((theme) => ({
   },
   deliveryDetails: {
     marginTop: theme.spacing(0.5),
-  },
-  unassignButton: {
-    color: color.darkPink,
-    textDecoration: "underline",
   },
 }));
 
@@ -81,40 +67,7 @@ const MissionDetailsStatus = ({ status, volunteerName }) => {
   );
 };
 
-const MissionDetailsIconList = ({ contentItems, classes, outerClass }) => (
-  <Grid container className={outerClass} alignItems="flex-start">
-    {contentItems.map((contentItem, index) => {
-      const Icon = contentItem.icon;
-      const avatarImage = contentItem.avatar?.image;
-      const content = contentItem.content.map((content, index) => {
-        return (
-          <Body2
-            key={`content-item-txt-${index + 1}`}
-            color="textPrimary"
-            style={content.style}
-            className={classes.contentTypography}
-          >
-            {content.text}
-          </Body2>
-        );
-      });
-
-      return (
-        <React.Fragment key={`content-item-${index + 1}`}>
-          <Grid item xs={1}>
-            {Icon && <Icon color="primary" />}
-            {avatarImage && <Avatar className={classes.avatar} alt="Volunteer" src={avatarImage} />}
-          </Grid>
-          <Grid item xs={11}>
-            {content}
-          </Grid>
-        </React.Fragment>
-      );
-    })}
-  </Grid>
-);
-
-const MissionDetailsContent = ({ description, classes }) => (
+const MissionDetailsType = ({ description, classes }) => (
   <Box>
     <H5 align="left" color="textSecondary">
       Mission Type
@@ -124,62 +77,6 @@ const MissionDetailsContent = ({ description, classes }) => (
     </Body2>
   </Box>
 );
-
-const MissionDetailsUnassignMeButton = ({ status, unassignFromMission, classes }) =>
-  status === MissionStatus.tentative ||
-  (status === MissionStatus.assigned && (
-    <Button className={classes.unassignButton} disableElevation onClick={unassignFromMission}>
-      Unassign Me
-    </Button>
-  ));
-
-const MissionDetailsButton = ({
-  type,
-  volunteerForMission,
-  startMission,
-  markMissionAsDelivered,
-}) => {
-  switch (type) {
-    case MissionStatus.assigned:
-      return (
-        <StyledButton
-          color="primary"
-          variant="contained"
-          disableElevation
-          onClick={volunteerForMission}
-        >
-          Accept Mission
-        </StyledButton>
-      );
-    case MissionStatus.tentative:
-    case MissionStatus.assigned:
-      return (
-        <StyledButton color="primary" variant="contained" disableElevation onClick={startMission}>
-          Start Mission
-        </StyledButton>
-      );
-    case MissionStatus.started:
-      return (
-        <StyledButton
-          color="primary"
-          variant="contained"
-          disableElevation
-          onClick={markMissionAsDelivered}
-        >
-          Mark Mission as Delivered
-        </StyledButton>
-      );
-    case MissionStatus.delivered:
-    case MissionStatus.done:
-      return null;
-    default:
-      return (
-        <StyledButton color="primary" variant="contained" disableElevation>
-          Loading..
-        </StyledButton>
-      );
-  }
-};
 
 const MissionDetailsPickUpDeliveryHeader = ({ header, classes }) => (
   <Body2 align="left" className={classes.deliveryDetailsHeader} color="textPrimary">
@@ -269,47 +166,38 @@ const MissionDetailsCard = ({
           title={mission.title}
           titleTypographyProps={{ variant: "h3", component: "span", color: "textPrimary" }}
           subheader={
-            <MissionDetailsIconList
-              outerClass={classes.subheader}
-              contentItems={subheaderItems}
-              classes={classes}
-            />
+            <MissionDetailsIconList outerClass={classes.subheader} contentItems={subheaderItems} />
           }
           className={classes.cardHeader}
         />
         <CardMedia image={cameraImage} title="Mission image" className={classes.image} />
         <CardContent className={classes.cardContent}>
-          <MissionDetailsContent description={mission.description} classes={classes} />
+          <MissionDetailsType description={mission.description} classes={classes} />
           <MissionDetailsPickUpDeliveryHeader header="Pick Up Details" classes={classes} />
           <MissionDetailsIconList
             outerClass={classes.deliveryDetails}
             contentItems={pickUpDetails}
-            classes={classes}
           />
           <MissionDetailsPickUpDeliveryHeader header="Delivery Details" classes={classes} />
           <MissionDetailsIconList
             outerClass={classes.deliveryDetails}
             contentItems={deliveryDetails}
-            classes={classes}
           />
         </CardContent>
         <Divider />
-        <CardActions>
+        <CardActions className={classes.cardAction}>
           <MissionDetailsButton
-            type={mission.status}
+            status={mission.status}
             volunteerForMission={() => volunteerForMission(mission.id)}
             startMission={() => startMission(mission.id)}
             markMissionAsDelivered={() => markMissionAsDelivered(mission.id)}
           />
         </CardActions>
-        <CardActions>
-          <Grid container justify="center">
-            <MissionDetailsUnassignMeButton
-              status={mission.status}
-              unassignFromMission={unassignFromMission}
-              classes={classes}
-            />
-          </Grid>
+        <CardActions className={classes.cardAction}>
+          <MissionDetailsUnassignMeButton
+            status={mission.status}
+            unassignFromMission={unassignFromMission}
+          />
         </CardActions>
       </Card>
       <UserPhoneUnverifiedPopup
