+++ src/app/component/MissionDetailsCard/index.jsx
@@ -1,0 +1,2 @@
+import MissionDetailsCard from "./MissionDetailsCard";
+export default MissionDetailsCard;
