+++ src/app/component/index.js
@@ -2,7 +2,6 @@ import Button from "./Button";
 import Chip from "./Chip";
 import MissionCard from "./MissionCard";
 import MissionList from "./MissionList";
-import MissionDetailsCard from "./MissionDetailsCard";
 import PrivateComponent from "./PrivateComponent";
 import ProtectedRoute from "./PrivateComponent";
 import Popup from "./Popup";
@@ -14,7 +13,6 @@ export {
   Chip,
   MissionCard,
   MissionList,
-  MissionDetailsCard,
   PrivateComponent,
   ProtectedRoute,
   Popup,
